Lab5/pzpi-22-6-sharaiev-bohdan-lab5/RadGuard/requirements.txt
asgiref==3.8.1
attrs==24.2.0
chardet==5.2.0
Django==5.1.4
django-cors-headers==4.6.0
django-mssql-backend==2.8.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docutils==0.21.2
drf-spectacular==0.28.0
drf-yasg==1.21.8
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
numpy==2.2.0
packaging==24.2
pillow==11.0.0
psycopg2==2.9.10
PyJWT==2.10.1
pyodbc==5.2.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
reportlab==4.2.5
rpds-py==0.22.3
sql_server.pyodbc==1.0
sqlparse==0.5.3
statistic==0.1.6
statistics==1.0.3.5
tzdata==2024.2
uritemplate==4.1.1
dj_database_url
gunicorn